requirements.txt
attrs==21.4.0
beautifulsoup4==4.10.0

certifi==2021.10.8
charset-normalizer
click
Flask
Flask-SQLAlchemy==2.5.1
greenlet
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous
Jinja2
MarkupSafe
packaging==21.3
pluggy==1.0.0
psycopg2-binary
py==1.11.0
pyparsing==3.0.7
requests==2.27.1
soupsieve==2.3.1
tomli
urllib3
Werkzeug
zipp==3.7.0

colorama~=0.4.4
pip~=21.2.2
wheel~=0.37.1
six~=1.16.0
setuptools~=58.0.4
fbprophet~=0.7.1